Changed ids from long to BigInteger.
This is to avoid issues between ids passed from ruby where they are
unsigned 64 bit ints to java where they are signed longs (also 64 bits).

+++ telemetry-lib/src/main/java/com/yammer/telemetry/tracing/InMemorySpanSinkSource.java
@@ -1,11 +1,12 @@
 package com.yammer.telemetry.tracing;
 
+import java.math.BigInteger;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
 public class InMemorySpanSinkSource implements SpanSink, SpanSource {
-    private final ConcurrentMap<Long, Trace> traces = new ConcurrentHashMap<>();
+    private final ConcurrentMap<BigInteger, Trace> traces = new ConcurrentHashMap<>();
 
     @Override
     public Collection<Trace> getTraces() {
@@ -13,7 +14,7 @@ public class InMemorySpanSinkSource implements SpanSink, SpanSource {
     }
 
     @Override
-    public Trace getTrace(long traceId) {
+    public Trace getTrace(BigInteger traceId) {
         return traces.get(traceId);
     }
 
@@ -32,7 +33,7 @@ public class InMemorySpanSinkSource implements SpanSink, SpanSource {
     }
 
     @Override
-    public void recordAnnotation(long traceId, long spanId, AnnotationData annotationData) {
+    public void recordAnnotation(BigInteger traceId, BigInteger spanId, AnnotationData annotationData) {
         final Trace newTrace = new Trace(traceId);
         newTrace.addAnnotation(spanId, annotationData);
