+++ telemetry-service/src/main/java/com/yammer/telemetry/service/api/TreeSpan.java
@@ -4,17 +4,18 @@ import com.google.common.collect.ImmutableList;
 import com.yammer.telemetry.tracing.AnnotationData;
 import org.joda.time.DateTime;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public class TreeSpan {
-    private final long id;
+    private final BigInteger id;
     private final String name;
     private final DateTime startTimeNanos;
     private final long duration;
     private final List<AnnotationData> annotations;
     private final ImmutableList<TreeSpan> children;
 
-    public TreeSpan(long id, String name, DateTime startTimeNanos, long duration, List<AnnotationData> annotations, ImmutableList<TreeSpan> children) {
+    public TreeSpan(BigInteger id, String name, DateTime startTimeNanos, long duration, List<AnnotationData> annotations, ImmutableList<TreeSpan> children) {
         this.id = id;
         this.name = name;
         this.startTimeNanos = startTimeNanos;
@@ -23,7 +24,7 @@ public class TreeSpan {
         this.children = children;
     }
 
-    public long getId() {
+    public BigInteger getId() {
         return id;
     }
